requirements.txt
pandas
biopython
beautifulsoup4
requests
streamlit
openpyxl
